backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pgeocode==0.5.0
alembic==1.14.1
reportlab==4.1.0
stripe==7.0.0
openai>=1.60.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
requests==2.32.3
